Moving shared functions between recipient views to helper file, added duns to reshape filters
