+++ usaspending_api/recipient/v2/views/states.py
@@ -1,9 +1,7 @@
 import logging
 from collections import OrderedDict
 from datetime import datetime
-from dateutil.relativedelta import relativedelta
 
-from django.conf import settings
 from django.contrib.postgres.aggregates import StringAgg
 from django.db.models import Sum
 from rest_framework.response import Response
@@ -16,6 +14,7 @@ from usaspending_api.common.exceptions import InvalidParameterException
 from usaspending_api.common.helpers.generic_helper import generate_fiscal_year
 from usaspending_api.common.views import APIDocumentationView
 from usaspending_api.recipient.models import StateData
+from usaspending_api.recipient.v2.helpers import validate_year, reshape_filters
 
 logger = logging.getLogger(__name__)
 
@@ -40,45 +39,6 @@ def validate_fips(fips):
         raise InvalidParameterException('Invalid fips: {}.'.format(fips))
 
     return fips
-
-
-def validate_year(year=None):
-    if year and not (year.isdigit() or year in ['all', 'latest']):
-        raise InvalidParameterException('Invalid year: {}.'.format(year))
-    return year
-
-
-def reshape_filters(state_code=None, year=None, award_type_codes=None):
-    # recreate filters
-    filters = {}
-
-    if state_code:
-        filters['place_of_performance_locations'] = [{'country': 'USA', 'state': state_code}]
-
-    if year:
-        today = datetime.now()
-        if year and year.isdigit():
-            time_period = [{
-                'start_date': '{}-10-01'.format(int(year) - 1),
-                'end_date': '{}-09-30'.format(year)
-            }]
-        elif year == 'all':
-            time_period = [{
-                'start_date': settings.API_SEARCH_MIN_DATE,
-                'end_date': datetime.strftime(today, '%Y-%m-%d')
-            }]
-        else:
-            last_year = today - relativedelta(years=1)
-            time_period = [{
-                'start_date': datetime.strftime(last_year, '%Y-%m-%d'),
-                'end_date': datetime.strftime(today, '%Y-%m-%d')
-            }]
-        filters['time_period'] = time_period
-
-    if award_type_codes:
-        filters['award_type_codes'] = award_type_codes
-
-    return filters
 
 
 def obtain_state_totals(fips, year=None, award_type_codes=None, subawards=False):
